Added prettier rule so that all json files are formatted with the json parser
For some strange reason, while I was setting up the eslint config, prettier kept adding trailing commas which was causing problems. It didn't do it for any of the other json files like package.json or .vscode/settings.json, but it did for eslintrc.json. Creating an override for *.json files to use the json parser option fixed it though. Still it feels a bit hacky for something that shouldn't even be a problem.