Added pre-commit hook for linting files
Installed husky which is a git hook manager and gave it the precommit hook of running the lint script so that it blocks the commit if the linter detects any issues. The problem with this is that it runs the linter on all files not just the ones that are to be committed.